create 3 slaves in storeShoppingCenter: viewStoreInfo, ViewStoreStock, AddToCard and OpenStore.
TODO: learn to count...
