+++ SadnaSrc/SadnaSrc/StoreCenter/StoreShoppingService.cs
@@ -29,18 +29,12 @@ namespace SadnaSrc.StoreCenter
         }
         public MarketAnswer OpenStore(string storeName, string address)
         {
-            MarketLog.Log("StoreCenter", "trying to add new store");
             try
             {
-                _shopper.ValidateRegistered();
-                MarketLog.Log("StoreCenter", "premission gained");
-                Store newStore = new Store(storeLogic.GetNextStoreId(), storeName, address);
-                storeLogic.AddStore(newStore);
-                MarketLog.Log("StoreCenter", "store was opened");
-                _shopper.AddOwnership(storeName);
-                stores.AddLast(newStore);
-                MarketLog.Log("StoreCenter", "add myself to the store list");
-                return new StoreAnswer(OpenStoreStatus.Success, "Store " + storeName + " has been opened successfully");
+                OpenStoreSlave slave = new OpenStoreSlave(_shopper);
+                Store S = slave.OpenStore(storeName, address);
+                stores.AddLast(S);
+                return slave.answer;
             }
             catch (StoreException e)
             {
@@ -55,17 +49,14 @@ namespace SadnaSrc.StoreCenter
                     "User validation as store owner has been failed. only registered users can open new stores. Error message has been created!");
             }
         }
-
+        
         public MarketAnswer ViewStoreInfo(string store)
         {
-            MarketLog.Log("StoreCenter", "");
             try
             {
-                _shopper.ValidateCanBrowseMarket();
-                MarketLog.Log("StoreCenter", "");
-                string[] storeInfo = storeLogic.GetStoreInfo(store);
-                MarketLog.Log("StoreCenter", "");
-                return new StoreAnswer(ViewStoreStatus.Success, "Store info has been successfully granted!", storeInfo);
+                ViewStoreInfoSlave slave = new ViewStoreInfoSlave(_shopper);
+                slave.ViewStoreInfo(store);
+                return slave.answer;
             }
             catch (StoreException e)
             {
@@ -79,47 +70,18 @@ namespace SadnaSrc.StoreCenter
                 return new StoreAnswer(ViewStoreStatus.InvalidUser,
                     "User validation as valid customer has been failed . only valid users can browse market. Error message has been created!");
             }
+
+
         }
-        private string GetProductStockInformation(string ProductID)
-        {
-            ModuleGlobalHandler handler = ModuleGlobalHandler.GetInstance();
-            StockListItem stockListItem = handler.DataLayer.GetStockListItembyProductID(ProductID);
-            if (stockListItem == null)
-            {
-                MarketLog.Log("storeCenter", "product not exists");
-                throw new StoreException(StoreEnum.ProductNotFound, "product " + ProductID + " does not exist in Stock");
-            }
-            string discount = "";
-            string product = stockListItem.Product.ToString();
-            if (stockListItem.Discount != null)
-                discount = stockListItem.Discount.ToString() + " , ";
-            string purchaseWay = handler.PrintEnum(stockListItem.PurchaseWay);
-            string quanitity = stockListItem.Quantity + "";
-            string result = product + " , " + discount + purchaseWay + " , " + quanitity;
-            return result;
-        }
+     
+        
         public MarketAnswer ViewStoreStock(string storename)
         {
-            MarketLog.Log("StoreCenter", "checking store stack");
             try
             {
-                _shopper.ValidateCanBrowseMarket();
-                MarketLog.Log("StoreCenter", "check if store exists");
-                if (!storeLogic.DataLayer.IsStoreExistAndActive(storename))
-                {
-                    MarketLog.Log("StoreCenter", "store do not exists");
-                    throw new StoreException(StoreEnum.StoreNotExists, "store not exists or active");
-                }
-                Store store = storeLogic.DataLayer.getStorebyName(storename);
-                LinkedList<string> result = new LinkedList<string>();
-                LinkedList<string> IDS = storeLogic.DataLayer.GetAllStoreProductsID(store.SystemId);
-                foreach (string item in IDS)
-                {
-                    result.AddLast(GetProductStockInformation(item));
-                }
-                string[] resultArray = new string[result.Count];
-                result.CopyTo(resultArray, 0);
-                return new StoreAnswer(StoreEnum.Success, "", resultArray);
+                ViewStoreStockSlave slave = new ViewStoreStockSlave(_shopper);
+                slave.ViewStoreStock(storename);
+                return slave.answer;
             }
             catch (StoreException e)
             {
@@ -135,44 +97,11 @@ namespace SadnaSrc.StoreCenter
 
         public MarketAnswer AddProductToCart(string store, string productName, int quantity)
         {
-            MarketLog.Log("StoreCenter", "trying to add something to the cart");
-
             try
             {
-                MarketLog.Log("StoreCenter", "checking if store exists");
-                if (!storeLogic.DataLayer.IsStoreExistAndActive(store))
-                { throw new StoreException(StoreEnum.StoreNotExists, "store not exists or active"); }
-                MarketLog.Log("StoreCenter", "checking if user has premmisions");
-                _shopper.ValidateCanBrowseMarket();
-                MarketLog.Log("StoreCenter", "checking if product exists");
-                if (storeLogic.IsProductNameAvailableInStore(store, productName)) //aka product is NotFiniteNumberException in store
-                {
-                    MarketLog.Log("StoreCenter", "Product is not exists in the store");
-                    throw new StoreException(StoreEnum.ProductNotFound, "product is not exists");
-                }
-                StockListItem stockListItem = storeLogic.GetProductFromStore(store, productName);
-                MarketLog.Log("StoreCenter", "checking that the required quantity is not too big");
-                if (quantity > stockListItem.Quantity)
-                {
-                    MarketLog.Log("StoreCenter", "required quantity is not too big");
-                    throw new StoreException(StoreEnum.QuantityIsTooBig, "required quantity is not too big");
-                }
-                MarketLog.Log("StoreCenter", "checking that the required quantity is not negative or zero");
-                if (quantity <= 0)
-                {
-                    MarketLog.Log("StoreCenter", "required quantity is negative or zero");
-                    throw new StoreException(StoreEnum.quantityIsNegatie, "required quantity is negative");
-                }
-                if (stockListItem.Discount != null)
-                {
-                    if (stockListItem.Discount.discountType == discountTypeEnum.Visible)
-                        if (stockListItem.Discount.checkTime())
-                            stockListItem.Product.BasePrice = stockListItem.Discount.CalcDiscount(stockListItem.Product.BasePrice);
-                }
-                _shopper.AddToCart(stockListItem.Product, store, quantity);
-                MarketLog.Log("StoreCenter", "add product successeded");
-                return new StoreAnswer(StoreEnum.Success, quantity + " " + productName + " from " + store + "has been" +
-                                                                 " successfully added to the user's cart!");
+                AddProductToCartSlave slave = new AddProductToCartSlave(_shopper);
+                slave.AddProductToCart(store, productName, quantity);
+                return slave.answer;
             }
             catch (StoreException e)
             {
@@ -203,3 +132,4 @@ namespace SadnaSrc.StoreCenter
         }
     }
 }
+
